backend/requirements.txt
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
markdown==3.7
djoser==2.1.0
webcolors==1.11.1
psycopg2-binary==2.9.9
Pillow==10.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0.2
gunicorn==20.1.0
python-dotenv==1.0.1